move stuff into better packages
