+++ fl119-0-1-2-compat/src/main/java/io/github/bumblesoftware/fastload/mixin/local/Hook1192.java
@@ -1,7 +1,7 @@
 package io.github.bumblesoftware.fastload.mixin.local;
 
-import io.github.bumblesoftware.fastload.abstraction.AbstractClientCalls;
-import io.github.bumblesoftware.fastload.abstraction.client119.Client119;
+import io.github.bumblesoftware.fastload.abstraction.tool.AbstractClientCalls;
+import io.github.bumblesoftware.fastload.abstraction.client.Client119;
 import io.github.bumblesoftware.fastload.config.init.FLMath;
 import io.github.bumblesoftware.fastload.init.Fastload;
 import io.github.bumblesoftware.fastload.init.FastloadClient;
