+++ fl1193compat/src/main/java/io/github/bumblesoftware/fastload/abstraction/client/Client1193.java
@@ -1,33 +1,31 @@
-package io.github.bumblesoftware.fastload.abstraction.client119;
+package io.github.bumblesoftware.fastload.abstraction.client;
+
 
 import com.mojang.serialization.Codec;
-import io.github.bumblesoftware.fastload.abstraction.RetrieveValueFunction;
-import io.github.bumblesoftware.fastload.abstraction.StoreValueFunction;
-import io.github.bumblesoftware.fastload.abstraction.client1182.Client1182;
-import io.github.bumblesoftware.fastload.config.screen.FLConfigScreen119;
+import io.github.bumblesoftware.fastload.abstraction.tool.RetrieveValueFunction;
+import io.github.bumblesoftware.fastload.abstraction.tool.StoreValueFunction;
 import io.github.bumblesoftware.fastload.util.MinMaxHolder;
-import net.minecraft.client.gui.screen.Screen;
+import net.minecraft.client.gui.widget.ButtonWidget;
 import net.minecraft.client.option.GameOptions;
 import net.minecraft.client.option.SimpleOption;
 import net.minecraft.text.Text;
 
-public class Client119 extends Client1182 {
-    @Override
-    public Screen newFastloadConfigScreen(final Screen parent) {
-        return new FLConfigScreen119(parent);
-    }
+public class Client1193 extends Client119 {
 
     @Override
-    public Text newTranslatableText(final String content) {
-        return Text.translatable(content);
+    public ButtonWidget getNewButton(
+            final int x,
+            final int y,
+            final int width,
+            final int height,
+            final Text message,
+            final ButtonWidget.PressAction onPress
+    ) {
+        return ButtonWidget.builder(message, onPress)
+                .dimensions(x, y, width, height)
+                .build();
     }
 
-    @Override
-    public Text newLiteralText(final String content) {
-        return Text.literal(content);
-    }
-
-    @SuppressWarnings("unchecked")
     @Override
     public SimpleOption<Boolean> newCyclingButton(
             final String namespace,
@@ -43,7 +41,6 @@ public class Client119 extends Client1182 {
         );
     }
 
-    @SuppressWarnings("unchecked")
     @Override
     public SimpleOption<Integer> newSlider(
             final String namespace,
@@ -72,10 +69,5 @@ public class Client119 extends Client1182 {
                 Integer.parseInt(retrieveValueFunction.getValue(identifier)),
                 value -> storeValueFunction.setValue(identifier, Integer.toString(value))
         );
-    }
-
-    @Override
-    public boolean isWindowFocused() {
-        return getClientInstance().isWindowFocused();
     }
 }
