+++ fl1194compat/src/main/java/io/github/bumblesoftware/fastload/abstraction/client/Client1194.java
@@ -1,7 +1,6 @@
-package io.github.bumblesoftware.fastload.abstraction.client1194;
+package io.github.bumblesoftware.fastload.abstraction.client;
 
 
-import io.github.bumblesoftware.fastload.abstraction.client1193.Client1193;
 import net.minecraft.client.font.TextRenderer;
 import net.minecraft.client.gui.DrawableHelper;
 import net.minecraft.client.gui.screen.Screen;
