+++ src/main/java/io/github/bumblesoftware/fastload/abstraction/tool/RetrieveValueFunction.java
@@ -1,4 +1,4 @@
-package io.github.bumblesoftware.fastload.abstraction;
+package io.github.bumblesoftware.fastload.abstraction.tool;
 
 public interface RetrieveValueFunction {
     String getValue(String key);
