+++ src/main/java/io/github/bumblesoftware/fastload/abstraction/tool/StoreValueFunction.java
@@ -1,4 +1,4 @@
-package io.github.bumblesoftware.fastload.abstraction;
+package io.github.bumblesoftware.fastload.abstraction.tool;
 
 public interface StoreValueFunction {
     void setValue(String key, String value);
